backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet>=3.0.0
openai==1.51.0
httpx==0.27.2
pandas==2.2.3
openpyxl==3.1.5
PyPDF2==3.0.1
pdfplumber==0.11.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
beautifulsoup4==4.12.3
